Restructured package
The Logger and Updater classes are now included in a models subpackage. Entry points are contained in cmd.py, while config management has been moved to cfg.py and utility.py has been renamed to helpers.py.
