+++ pipupdater/models/Logger.py
@@ -15,6 +15,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
+
 import smooth_logger
 
 from os import makedirs
