+++ pipupdater/models/__init__.py
@@ -16,13 +16,5 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 
-from .cmd import entry_point
-
-from .cfg import get_args
-from .cfg import get_config
-from .cfg import VERSION
-
-from .helpers import str_starts_with
-
-from .models import Logger
-from .models import Updater
+from .Logger import Logger
+from .Updater import Updater
